requirements/base.txt
attrs==24.2.0
bcrypt==4.2.0
boto3==1.35.12
celery==5.4.0
django==4.2.9
django-celery-beat==2.7.0
django-celery-results==2.5.1
flower==2.0.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-jwt==1.19.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
psycopg2-binary